refactor(pass-7): extract freshness stamping core wrapper
