+++ src/core/index.ts
@@ -1,0 +1,3 @@
+export { LAMBDA, calculateFreshnessScore, scoreLabel } from "./decay.js";
+export { looksLikeFailedAdapterContent } from "./guards.js";
+export { stampFreshness, toStructuredJSON, formatForLLM } from "./envelope.js";
